added authMiddleware to protect some routes

+++ backend/controllers/notesController.js
@@ -2,15 +2,16 @@ const Note = require('../models/noteModels');
 const asyncHandler = require('express-async-handler');
 
 
+
 const getNotes =asyncHandler(async (req,res)=>{
-    const notes = await Note.find();
+    const notes = await Note.find({user:req.user.id});
         res.status(200).json(notes)
     
 })
 
 const addNote =asyncHandler(async (req,res) =>{
     const note = await Note.create({
-        title:req.body.title,body:req.body.body
+        title:req.body.title,body:req.body.body,user:req.user.id
     })
     res.status(200).json(note);
 })
@@ -25,6 +26,20 @@ const updateNote =asyncHandler(async (req,res) =>{
         throw new Error("No note found with that Id")
     } 
 
+    const user = await User.findById(req.user.id);
+
+    if(!user){
+        res.status(401)
+
+        throw new Error("NO user found with this Id")
+    }
+
+    if(note.user.toString() !== user.id){
+        res.status(401)
+
+        throw new Error("User is not authorized to update")
+    }
+
     const updatedNote = await Note.findByIdAndUpdate(req.params.id,req.body,{new:true});    
     res.json(updatedNote)
 })
@@ -38,6 +53,20 @@ const deleteNote =asyncHandler(async (req,res) =>{
         res.status(400)
 
         throw new Error("No note found with that id")
+    }
+
+    const user = await User.findById(req.user.id);
+
+    if(!user){
+        res.status(401)
+
+        throw new Error("NO user found with this Id")
+    }
+
+    if(note.user.toString() !== user.id){
+        res.status(401)
+
+        throw new Error("User is not authorized to delete");
     }
 
     await Note.findByIdAndDelete(req.params.id)
